requirements.txt
stellar-sdk>=9.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.0.0
